+++ app/include/tab/dashboard.hpp
@@ -8,6 +8,7 @@
 
 class AutoTabFrame;
 class RecyclingGrid;
+class HRecyclerFrame;
 
 class Dashboard : public brls::Box {
 public:
@@ -19,14 +20,11 @@ private:
     BRLS_BIND(brls::Label, labelServer, "dashboard/server/version");
     BRLS_BIND(brls::Label, labelName, "dashboard/server/name");
     BRLS_BIND(brls::Label, labelAddr, "dashboard/server/addr");
-    BRLS_BIND(brls::Label, labelMovie, "dashboard/movie/count");
-    BRLS_BIND(brls::Label, labelSeries, "dashboard/series/count");
-    BRLS_BIND(brls::Label, labelEpisode, "dashboard/episode/count");
-    BRLS_BIND(brls::Label, labelSong, "dashboard/song/count");
     BRLS_BIND(brls::Button, btnRestart, "dashboard/restart");
     BRLS_BIND(brls::Button, btnScan, "dashboard/scan");
     BRLS_BIND(RecyclingGrid, activity, "dashboard/activity");
     BRLS_BIND(RecyclingGrid, sess, "dashboard/session");
+    BRLS_BIND(HRecyclerFrame, itemCount, "dashboard/count");
     BRLS_BIND(brls::Box, storage, "dashboard/storage/box");
 
     std::unordered_map<std::string, std::string> taskMap;