+++ app/src/main.cpp
@@ -68,6 +68,8 @@ int main(int argc, char* argv[]) {
         return EXIT_FAILURE;
     }
 
+    conf.initThemes();
+
     // Return directly to the desktop when closing the application (only for NX)
     brls::Application::getPlatform()->exitToHomeMode(true);
 
@@ -95,27 +97,6 @@ int main(int argc, char* argv[]) {
     brls::Application::registerXMLView("SearchTab", SearchTab::create);
     brls::Application::registerXMLView("RemoteTab", RemoteTab::create);
     brls::Application::registerXMLView("SettingTab", SettingTab::create);
-
-    brls::Theme::getLightTheme().addColor("color/app", nvgRGB(2, 176, 183));
-    brls::Theme::getDarkTheme().addColor("color/app", nvgRGB(51, 186, 227));
-    // 用于骨架屏背景色
-    brls::Theme::getLightTheme().addColor("color/grey_1", nvgRGB(245, 246, 247));
-    brls::Theme::getDarkTheme().addColor("color/grey_1", nvgRGB(51, 52, 53));
-    brls::Theme::getLightTheme().addColor("color/grey_2", nvgRGB(245, 245, 245));
-    brls::Theme::getDarkTheme().addColor("color/grey_2", nvgRGB(51, 53, 55));
-    brls::Theme::getLightTheme().addColor("color/grey_3", nvgRGBA(200, 200, 200, 16));
-    brls::Theme::getDarkTheme().addColor("color/grey_3", nvgRGBA(160, 160, 160, 160));
-    brls::Theme::getLightTheme().addColor("color/danger", nvgRGB(198, 28, 28));
-    brls::Theme::getDarkTheme().addColor("color/danger", nvgRGB(198, 28, 28));
-    // 分割线颜色
-    brls::Theme::getLightTheme().addColor("color/line", nvgRGB(208, 208, 208));
-    brls::Theme::getDarkTheme().addColor("color/line", nvgRGB(100, 100, 100));
-    // 深浅配色通用的灰色字体颜色
-    brls::Theme::getLightTheme().addColor("font/grey", nvgRGB(148, 153, 160));
-    brls::Theme::getDarkTheme().addColor("font/grey", nvgRGB(148, 153, 160));
-
-    brls::getStyle().addMetric("main/content_padding_sides", 25.f);
-    brls::getStyle().addMetric("main/content_padding_top_bottom", 30.f);
 
     if (!brls::Application::getPlatform()->isApplicationMode()) {
         brls::Application::pushActivity(new HintActivity());