+++ app/src/tab/dashboard.cpp
@@ -24,6 +24,61 @@ inline const brls::ButtonStyle BUTTONSTYLE_DANDER = {
     .disabledBorderColor = "",
 };
 
+// ----------------- Media Count ---------------------------
+
+using ItemCount = std::pair<std::string, long>;
+
+class ItemCountCell : public RecyclingGridItem {
+public:
+    ItemCountCell() {
+        this->setPadding(15);
+        this->setAlignItems(brls::AlignItems::CENTER);
+        this->setJustifyContent(brls::JustifyContent::SPACE_BETWEEN);
+        this->setBackgroundColor(brls::Application::getTheme()["color/grey_2"]);
+
+        this->name = new brls::Label();
+        this->count = new brls::Label();
+
+        this->addView(this->name);
+        this->addView(this->count);
+    }
+
+    void setCell(const ItemCount& it) {
+        this->name->setText(it.first);
+        this->count->setText(std::to_string(it.second));
+    }
+
+private:
+    brls::Label* name;
+    brls::Label* count;
+};
+
+class ItemCountDataSource : public RecyclingGridDataSource {
+public:
+    using ItemList = std::vector<ItemCount>;
+
+    explicit ItemCountDataSource(const jellyfin::ItemCount& r) {
+        this->list.push_back({"main/search/genres/movie"_i18n, r.MovieCount});
+        this->list.push_back({"main/search/genres/series"_i18n, r.SeriesCount});
+        this->list.push_back({"main/search/genres/music"_i18n, r.SongCount});
+        this->list.push_back({"main/search/genres/episode"_i18n, r.EpisodeCount});
+    }
+
+    size_t getItemCount() override { return this->list.size(); }
+
+    RecyclingGridItem* cellForRow(RecyclingView* recycler, size_t index) override {
+        ItemCountCell* cell = dynamic_cast<ItemCountCell*>(recycler->dequeueReusableCell("Cell"));
+        auto& item = this->list.at(index);
+        cell->setCell(item);
+        return cell;
+    }
+
+    void clearData() override { this->list.clear(); }
+
+private:
+    ItemList list;
+};
+
 // ----------------- Device Tab ----------------------------
 
 class DeviceCell : public RecyclingGridItem {
@@ -354,7 +409,14 @@ Dashboard::Dashboard() {
     this->tabFrame->registerTabAction(this);
     this->activity->registerCell("Cell", []() { return new ActivityLogCell("xml/view/activity_warn.xml"); });
     this->sess->registerCell("Cell", []() { return new SessionCell(); });
-    this->doItemCount();
+    this->itemCount->registerCell("Cell", []() { return new ItemCountCell(); });
+
+    if (brls::Application::ORIGINAL_WINDOW_HEIGHT < 720) {
+        this->itemCount->setVisibility(brls::Visibility::GONE);
+    } else {
+        this->doItemCount();
+    }
+
     this->doSystemInfo();
     this->doActivityWarn();
     this->doSession();
@@ -369,10 +431,7 @@ void Dashboard::doItemCount() {
     jellyfin::getJSON<jellyfin::ItemCount>(
         [ASYNC_TOKEN](const jellyfin::ItemCount& r) {
             ASYNC_RELEASE
-            this->labelMovie->setText(std::to_string(r.MovieCount));
-            this->labelSeries->setText(std::to_string(r.SeriesCount));
-            this->labelEpisode->setText(std::to_string(r.EpisodeCount));
-            this->labelSong->setText(std::to_string(r.SongCount));
+            this->itemCount->setDataSource(new ItemCountDataSource(r));
         },
         [ASYNC_TOKEN](const std::string& ex) {
             ASYNC_RELEASE
@@ -486,7 +545,7 @@ void Dashboard::doRunTask(const std::string& id) {
 }
 
 void Dashboard::doStorage() {
-    brls::View *cell = new SkeletonCell();
+    brls::View* cell = new SkeletonCell();
     cell->setGrow(1.0f);
     this->storage->addView(cell);