+++ app/src/tab/media_collection.cpp
@@ -61,7 +61,7 @@ public:
     GenresTab(const std::string& itemId, const std::string& itemType) {
         this->setGrow(1.f);
         this->registerCell("Cell", VideoCardCell::create);
-        this->spanCount = 6;
+        this->spanCount = brls::getStyle().getMetric("app/grid/6");
 
         std::string query = HTTP::encode_form({
             {"userId", AppConfig::instance().getUserId()},
@@ -90,7 +90,7 @@ public:
     ArtistsTab(const std::string& itemId) : itemId(itemId) {
         this->setGrow(1.f);
         this->registerCell("Cell", VideoCardCell::create);
-        this->spanCount = 6;
+        this->spanCount = brls::getStyle().getMetric("app/grid/6");
 
         this->onNextPage([this] { this->doRequest(); });
         this->doRequest();
@@ -177,7 +177,7 @@ MediaCollection::MediaCollection(const std::string& itemId, const std::string& i
 
     this->pageSize = this->recycler->spanCount * 3;
     if (itemType == jellyfin::mediaTypeMusicAlbum) {
-        this->recycler->estimatedRowHeight = 240;
+        this->recycler->estimatedRowHeight = 220;
     } else if (itemType == jellyfin::mediaTypeBook) {
         this->recycler->estimatedRowHeight = 280;
     }